requirements.txt
pyrogram==2.0.106
flask==3.1.1
python-dotenv==1.1.1
gunicorn==21.2.0